\name{AddSinVectors}
\alias{AddSinVectors}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
   Internal function.
}
\usage{
AddSinVectors(vectors)
}

\author{Ruiz-Cuetos, J.C., \email{bilba_t@hotmail.com}, Polo, M.E., \email{mepolo@unex.es}, Rodriguez, P.G. \email{pablogr@unex.es}}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
