\name{datos2.1}
\docType{data}
\alias{datos2.1}
\title{Data Set}
\description{
	Dataset obtained at UNEX Merida. Maria Eugenia Polo
}
\usage{datos2.1}
\format{A matrix containing multiples observations.}
\source{UNEX Merida}
\references{
	 Website \url{http://gim.unex.es/VecStatGraphs2D/}
}
\keyword{datasets}
