\name{DrawArc}
\alias{DrawArc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
   Internal function.
}
\usage{
DrawArc(x = 1, y = NULL, radius = 1, angle1 = deg1 * pi/180, angle2 = deg2 * pi/180, 
	deg1 = 0, deg2 = 45, n = 35, col = 1, ...)
}
%- maybe also 'usage' for other objects documented here.

\author{Ruiz-Cuetos, J.C., \email{jcruizcue@gmail.com}, Felicisimo, A.M., \email{amfeli@unex.es}, 
Cuartero, A., \email{acuartero@unex.es}, Polo, M.E., \email{mepolo@unex.es}, 
Rodriguez, P.G. \email{pablogr@unex.es}}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
