\name{RectangularData}
\docType{data}
\alias{RectangularData}
\title{GPS Data Set}
\description{
	Dataset obtained with a GPS, UNEX Merida.
}
\usage{RectangularData}
\format{A matrix containing 100 observations.}
\source{UNEX Merida}
\references{
	Escuela Universitaria Politecnica, Caceres.
}
\keyword{datasets}
