\name{VecStatGraphs2D}
\title{Vectorial analysis using graphical and analytical methods in 2D}
\description{This package performs a 2D statistical analysis, both numerical and graphical, of a set of vectors. Because a vector has two components, module and azimuth, vector analysis is performed in three stages. First, modules are analyzed using linear statistics. Second, azimuths are analyzed by circular statistics. Third a joint analysis of modules and azimuths is performed using density maps that allow detection of other distribution properties (i.e. anisotropy) and outliers. Tests and circular statistic parameters are accompanied by a full range of graphing: histograms, maps of distributions, point maps, vector maps, density maps, distribution modules and azimuths.}
\details{These are not to be called by the user.}
\keyword{internal}