\name{Histogram}
\alias{Histogram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   Internal function
}
\description{
   Internal function.
}
\usage{
Histogram(vectors, n_classes)
}
%- maybe also 'usage' for other objects documented here.

\author{Ruiz-Cuetos, J.C., \email{bilba_t@hotmail.com}, Polo, M.E., \email{mepolo@unex.es}, Rodriguez, P.G. \email{pablogr@unex.es}} 

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
