\name{MeasuredRealData}
\alias{MeasuredRealData}

\docType{data}
\title{GPS Data Set}
\description{
	Dataset obtained with a GPS, UNEX Merida.
}
\usage{MeasuredRealData}
\format{A matrix containing 53 observations. }

\references{
Website \url{http://gim.unex.es/VecStatGraphs2D/}
}
\keyword{datasets}