\name{RectangularData}
\docType{data}
\alias{RectangularData}
\title{GPS Data Set}
\description{
	Dataset obtained with a GPS, UNEX Merida.
}
\usage{RectangularData}
\format{A matrix containing 100 observations.}
\source{UNEX Merida}
\references{
	 Website \url{http://gim.unex.es/VecStatGraphs2D/}
}
\keyword{datasets}